pallet_staking_async_parachain_runtime
Module xcm_config

Modules

bridging

All configuration related to bridging

Structs

AmbassadorEntities

AuthorizeAliasHoldReason

CheckingAccount

Collectives

DepositPerByte

DepositPerItem

FellowshipAdminBodyId

FellowshipEntities

ForeignAssetsPalletLocation

GeneralAdminBodyId

LocalPlurality

MaxAssetsIntoHolding

MaxInstructions

ParentOrParentsPlurality

PoolAssetsPalletLocation

RelayChainOrigin

RelayNetwork

RelayTreasuryLocation

RootLocation

StakingAdminBodyId

StakingPot

TreasurerBodyId

TreasuryAccount

TrustBackedAssetsPalletIndex

TrustBackedAssetsPalletLocation

UniquesPalletLocation

UniversalLocation

UniversalLocationNetworkId

WestendLocation

XcmAssetFeesReceiver

XcmBenchmarkHelper

Simple conversion of u32 into an AssetId for use in benchmarking.

XcmConfig

Type Aliases

AssetFeeAsExistentialDepositMultiplierFeeCharger

AssetTransactors

Means for transacting assets on this chain.

Barrier

FellowshipAdminToPlurality

Type to convert the FellowshipAdmin origin to a Plurality Location value.

ForeignAssetFeeAsExistentialDepositMultiplierFeeCharger

Multiplier used for dedicated TakeFirstAssetTrader with ForeignAssets instance.

ForeignAssetsConvertedConcreteId

AssetId/Balance converter for ForeignAssets.

ForeignFungiblesTransactor

Means for transacting foreign assets from different global consensus.

FungibleTransactor

Means for transacting the native currency on this chain.

FungiblesTransactor

Means for transacting assets besides the native currency on this chain.

GeneralAdminToPlurality

Type to convert the GeneralAdmin origin to a Plurality Location value.

LocalOriginToLocation

Local origins on this chain are allowed to dispatch XCM sends/executions.

LocalPalletOriginToLocation

Type to convert a pallet Origin type value into a Location value which represents an interior location of this chain for a destination chain.

LocationToAccountId

Type for specifying how a Location can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

PoolAssetsConvertedConcreteId

AssetId/Balance converter for PoolAssets.

PoolAssetsExchanger

Asset converter for pool assets. Used to convert one asset to another, when there is a pool available between the two. This type thus allows paying fees with any asset as long as there is a pool between said asset and the asset required for fee payment.

PoolFungiblesTransactor

Means for transacting asset conversion pool assets on this chain.

PriceForParentDelivery

StakingAdminToPlurality

Type to convert the StakingAdmin origin to a Plurality Location value.

TreasurerToPlurality

Type to convert the Treasurer origin to a Plurality Location value.

TrustBackedAssetsConvertedConcreteId

AssetId/Balance converter for TrustBackedAssets.

TrustedTeleporters

Cases where a remote origin is accepted as trusted Teleporter for a given asset:

UniquesConvertedConcreteId

Matcher for converting ClassId/InstanceId into a uniques asset.

UniquesTransactor

Means for transacting unique assets.

WaivedLocations

Locations that will not be charged fees in the executor, either execution or delivery. We only waive fees for system functions, which these locations represent.

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.